support fractional shares

+++ src/shares/views.py
@@ -37,6 +37,26 @@ class SharesListView(ListView):
         print(data)
         data['goal_name_mapping'] = get_all_goals_id_to_name_mapping()
         data['user_name_mapping'] = get_all_users()
+        total_investment = 0
+        latest_value = 0
+        as_on_date= None
+        total_gain = 0
+        share_objs = Share.objects.all()
+        for share_obj in share_objs:
+            if not share_obj.latest_value:
+                continue
+            if not as_on_date:
+                as_on_date = share_obj.as_on_date
+            else:
+                if as_on_date < share_obj.as_on_date:
+                    as_on_date = share_obj.as_on_date
+            latest_value += share_obj.latest_value
+            total_investment += share_obj.buy_value
+            total_gain += share_obj.gain
+        data['as_on_date'] = as_on_date
+        data['total_gain'] = total_gain
+        data['total_investment'] = total_investment
+        data['latest_value'] = latest_value
         return data
 
 class ShareTransactionsListView(ListView):
@@ -130,18 +150,20 @@ def upload_transactions(request):
     template = 'shares/upload_transactions.html'
     # https://www.youtube.com/watch?v=Zx09vcYq1oc&list=PLLxk3TkuAYnpm24Ma1XenNeq1oxxRcYFT
     if request.method == 'POST':
-        uploaded_file = request.FILES['document']
-        user = request.POST['user']
-        broker = request.POST.get('brokerControlSelect')
-        print(uploaded_file)
-        print(broker)
-        fs = FileSystemStorage()
-        file_locn = fs.save(uploaded_file.name, uploaded_file)
-        print(file_locn)
-        print(settings.MEDIA_ROOT)
-        full_file_path = settings.MEDIA_ROOT + '/' + file_locn
-        add_transactions(broker, user, full_file_path)
-        fs.delete(file_locn)
+        print(request.POST)
+        if True:
+            uploaded_file = request.FILES['document']
+            user = request.POST['user']
+            broker = request.POST.get('brokerControlSelect')
+            print(uploaded_file)
+            print(broker)
+            fs = FileSystemStorage()
+            file_locn = fs.save(uploaded_file.name, uploaded_file)
+            print(file_locn)
+            print(settings.MEDIA_ROOT)
+            full_file_path = settings.MEDIA_ROOT + '/' + file_locn
+            add_transactions(broker, user, full_file_path)
+            fs.delete(file_locn)
     users = get_all_users()
     context = {'users':users}
     return render(request, template, context)
@@ -149,21 +171,62 @@ def upload_transactions(request):
 def add_transaction(request):
     template = 'shares/add_transaction.html'
     if request.method == 'POST':
-        exchange = request.POST['exchange']
-        symbol = request.POST['symbol']
-        user = request.POST['user']
-        print('user is of type:',type(user))
-        trans_date = get_date_or_none_from_string(request.POST['trans_date'])
-        trans_type = request.POST['trans_type']
-        price = get_float_or_none_from_string(request.POST['price'])
-        quantity = get_int_or_none_from_string(request.POST['quantity'])
-        conversion_rate = get_float_or_none_from_string(request.POST['conversion_rate'])
-        trans_price = get_float_or_none_from_string(request.POST['trans_price'])
-        broker = request.POST['broker']
-        notes = request.POST['notes']
-        insert_trans_entry(exchange, symbol, user, trans_type, quantity, price, trans_date, notes, broker, conversion_rate, trans_price)
+        if "submit" in request.POST:
+            exchange = request.POST['exchange']
+            symbol = request.POST['symbol']
+            user = request.POST['user']
+            print('user is of type:',type(user))
+            trans_date = get_date_or_none_from_string(request.POST['trans_date'])
+            trans_type = request.POST['trans_type']
+            price = get_float_or_none_from_string(request.POST['price'])
+            quantity = get_float_or_none_from_string(request.POST['quantity'])
+            conversion_rate = get_float_or_none_from_string(request.POST['conversion_rate'])
+            trans_price = get_float_or_none_from_string(request.POST['trans_price'])
+            broker = request.POST['broker']
+            notes = request.POST['notes']
+            insert_trans_entry(exchange, symbol, user, trans_type, quantity, price, trans_date, notes, broker, conversion_rate, trans_price)
+        else:
+            print('fetching exchange price')
+            exchange = request.POST['exchange']
+            symbol = request.POST['symbol']
+            user = request.POST['user']
+            trans_date = get_date_or_none_from_string(request.POST['trans_date'])
+            exchange_rate = 1
+            if exchange == 'NASDAQ' or exchange == 'NYSE':
+                exchange_rate = get_forex_rate(trans_date, 'USD', 'INR')
+            users = get_all_users()
+            context = {'users':users, 'operation': 'Add Transaction', 'conversion_rate':exchange_rate,
+                        'trans_date':trans_date.strftime("%Y-%m-%d"), 'user':user, 'exchange':exchange, 'symbol':symbol}
+            return render(request, template, context)
+
     users = get_all_users()
-    context = {'users':users, 'operation': 'Add Transaction'}
+    context = {'users':users, 'operation': 'Add Transaction', 'conversion_rate':1}
+    return render(request, template, context)
+
+def update_transaction(request,id):
+    template = 'shares/update_transaction.html'
+    trans = Transactions.objects.get(id=id)
+
+    if request.method == 'POST':
+        trans.trans_date = get_date_or_none_from_string(request.POST['trans_date'])
+        trans.trans_type = request.POST['trans_type']
+        trans.quantity = get_float_or_none_from_string(request.POST['quantity'])
+        trans.conversion_rate = get_float_or_none_from_string(request.POST['conversion_rate'])
+        trans_price = get_float_or_none_from_string(request.POST['trans_price'])
+        trans.price = trans_price
+        if not trans_price:
+            trans_price = trans.trans_price*trans.quantity*trans.conversion_rate
+        trans.trans_price = trans_price
+        trans.broker = request.POST['broker']
+        trans.notes = request.POST['notes']
+        trans.save()
+        reconcile_share(trans.share)
+    context = {
+        'user':get_user_name_from_id(trans.share.user), 'operation': 'Update Transaction',
+        'exchange': trans.share.exchange, 'symbol':trans.share.symbol, 'trans_date':trans.trans_date.strftime("%Y-%m-%d"),
+        'price': trans.price, 'quantity':trans.quantity, 'conversion_rate':trans.conversion_rate,
+        'trans_price': trans.trans_price, 'broker':trans.broker, 'notes':trans.notes
+    }
     return render(request, template, context)
 
 def add_transactions(broker, user, full_file_path):
@@ -174,6 +237,31 @@ def add_transactions(broker, user, full_file_path):
             insert_trans_entry(
                 trans["exchange"], trans["symbol"], user, trans["type"], trans["quantity"], trans["price"], trans["date"], trans["notes"], 'ZERODHA')
 
+def reconcile_share(share_obj):
+    quantity = 0
+    buy_value = 0
+    buy_price = 0
+    sell_quantity = 0
+    sell_transactions = Transactions.objects.filter(share=share_obj, trans_type='Sell')
+    for trans in sell_transactions:
+        sell_quantity = sell_quantity + trans.quantity
+    
+    buy_transactions = Transactions.objects.filter(share=share_obj, trans_type='Buy').order_by('trans_date')
+    for trans in buy_transactions:
+        if sell_quantity > 0:
+            if sell_quantity > trans.quantity:
+                sell_quantity = sell_quantity - trans.quantity
+            else:
+                quantity = trans.quantity - sell_quantity
+                buy_value = quantity*trans.trans_price*trans.conversion_rate
+        else:
+            quantity = quantity + trans.quantity
+            buy_value = buy_value + trans.quantity*trans.trans_price*trans.conversion_rate
+    buy_price = buy_value/quantity
+    share_obj.quantity = quantity
+    share_obj.buy_value = buy_value
+    share_obj.buy_price = buy_price
+    share_obj.save()
 
 def insert_trans_entry(exchange, symbol, user, trans_type, quantity, price, date, notes, broker, conversion_rate=1, trans_price=None):
     share_obj = None
@@ -201,14 +289,14 @@ def insert_trans_entry(exchange, symbol, user, trans_type, quantity, price, date
                                     broker=broker,
                                     notes=notes)
         if trans_type == 'Buy':
-            new_qty = int(share_obj.quantity)+quantity
+            new_qty = float(share_obj.quantity)+quantity
             new_buy_value = float(share_obj.buy_value) + trans_price
             share_obj.quantity = new_qty
             share_obj.buy_value = new_buy_value
             share_obj.buy_price = new_buy_value/float(new_qty)
             share_obj.save()
         else:
-            new_qty = int(share_obj.quantity)-quantity
+            new_qty = float(share_obj.quantity)-quantity
             if new_qty:
                 new_buy_value = float(share_obj.buy_value) - trans_price
                 share_obj.quantity = new_qty
@@ -216,7 +304,10 @@ def insert_trans_entry(exchange, symbol, user, trans_type, quantity, price, date
                 share_obj.buy_price = new_buy_value/float(new_qty)
                 share_obj.save()
             else:
-                share_obj.delete()
+                share_obj.quantity = 0
+                share_obj.buy_value = 0
+                share_obj.buy_price = 0
+                share_obj.save()
     except IntegrityError:
         print('Transaction exists')
 
@@ -226,20 +317,24 @@ def refresh(request):
     end = datetime.date.today()
     share_objs = Share.objects.all()
     for share_obj in share_objs:
-        if share_obj.as_on_date != datetime.date.today():
-            vals = get_latest_vals(share_obj.symbol, share_obj.exchange, start, end)
-            if vals:
-                for k, v in vals.items():
-                    if k and v:
-                        if not share_obj.as_on_date or k > share_obj.as_on_date:
-                            share_obj.as_on_date = k
-                            share_obj.latest_price = v
-                            if share_obj.exchange == 'NASDAQ':
-                                share_obj.conversion_rate = get_forex_rate(k, 'USD', 'INR')
-                            else:
-                                share_obj.conversion_rate = 1
-                            share_obj.latest_value = float(share_obj.latest_price) * float(share_obj.conversion_rate) * float(share_obj.quantity)
-                            share_obj.save()
+        latest_date = None
+        latest_val = None
+        vals = get_latest_vals(share_obj.symbol, share_obj.exchange, start, end)
+        if vals:
+            for k, v in vals.items():
+                if k and v:
+                    if not latest_date or k > latest_date:
+                        latest_date = k
+                        latest_val = v
+        if latest_date and latest_val:
+            share_obj.as_on_date = latest_date
+            if share_obj.exchange == 'NASDAQ':
+                share_obj.conversion_rate = get_forex_rate(k, 'USD', 'INR')
+            else:
+                share_obj.conversion_rate = 1
+            share_obj.latest_value = float(latest_val) * float(share_obj.conversion_rate) * float(share_obj.quantity)
+            share_obj.latest_price = latest_val * float(share_obj.conversion_rate)
+            share_obj.save()
         if share_obj.latest_value: 
             share_obj.gain=float(share_obj.latest_value)-float(share_obj.buy_value)
             share_obj.save()
@@ -274,3 +369,7 @@ class CurrentShares(APIView):
             data['as_on_date'] = share.as_on_date
             shares['shares'].append(data)
         return Response(shares)
+
+def delete_shares(request):
+    Share.objects.all().delete()
+    return HttpResponseRedirect('../')
